autodocs/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.12.1
billiard==4.2.4
blis==1.3.3
catalogue==2.0.10
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.13
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.128.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
kombu==5.6.2
Markdown==3.10
MarkupSafe==3.0.3
murmurhash==1.0.15
neo4j==6.1.0
numpy==2.4.1
openai==2.15.0
packaging==25.0
preshed==3.0.12
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
redis==7.1.0
requests==2.32.5
setuptools==80.9.0
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
srsly==2.5.2
starlette==0.50.0
thinc==8.3.10
tqdm==4.67.1
tree-sitter==0.25.2
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
wrapt==2.0.1
